[frontend] compute table width before calling dummy lines
